feat: 创建认证和数据集相关表结构，加载初始数据

+++ avalon-modules/avalon-authorization/src/main/java/io/github/lishangbu/avalon/authorization/repository/MenuRepository.java
@@ -16,7 +16,7 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface MenuRepository
-    extends ListCrudRepository<Menu, Long>, ListPagingAndSortingRepository<Menu, Long> {
+    extends ListCrudRepository<Menu, Integer>, ListPagingAndSortingRepository<Menu, Integer> {
   /**
    * 通过角色代码查询所有菜单
    *
